backend/requirements.txt
# Core FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
httpx>=0.25.2

# Data processing (Python 3.13 compatible)
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# AI/ML Dependencies (Python 3.13 compatible)
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.0
spacy>=3.7.0
tokenizers>=0.15.0
huggingface-hub>=0.19.0
safetensors>=0.4.0
datasets>=2.14.0

# NeMo Guardrails (may need manual install if not compatible)
# nemoguardrails>=0.8.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Build tools for Python 3.13
setuptools>=69.0.0
wheel>=0.42.0